+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog].
 
 ## [v1.0.11-1.16.5] - 2021-07-22
 ### Changed
+- Load configs during construct so they can be used in registry events
 - More minor log output changes
 
 ## [v1.0.10-1.16.5] - 2021-07-22